Add battery calibration support
This commit adds initial support for battery calibration. The
calibration mode can be enabled and disabled through sysfs, but the
WMI events that indicate the charging states and the completion of the
calibration are not being handled yet. Those events are logged as
unhandled WMI events with ids 8 (charging) and 11 (calibration status)
to the kernel log. After the calibration, the mode must be disabled
manually.